Export getUserSession into its own file
